+++ wolfhsm/wh_log_ringbuf.h
@@ -33,16 +33,16 @@
 
 /* Ring buffer configuration structure */
 typedef struct whLogRingbufConfig_t {
-    void*    buffer;      /* User-supplied buffer */
-    uint32_t buffer_size; /* Size of buffer in bytes */
+    void*  buffer;      /* User-supplied buffer */
+    size_t buffer_size; /* Size of buffer in bytes */
 } whLogRingbufConfig;
 
 /* Ring buffer context structure */
 typedef struct whLogRingbufContext_t {
     whLogEntry* entries;     /* Pointer to user-supplied buffer */
-    uint32_t    capacity;    /* Number of entries buffer can hold */
-    uint32_t    head;        /* Next write position (0 to capacity-1) */
-    uint32_t    count;       /* Current entries (0 to capacity) */
+    size_t      capacity;    /* Number of entries buffer can hold */
+    size_t      head;        /* Next write position (0 to capacity-1) */
+    size_t      count;       /* Current entries (0 to capacity) */
     int         initialized; /* Initialization flag */
 } whLogRingbufContext;
